WIP: lazy subdriver loading in driver
