+++ drivers/SmartThings/zwave-switch/src/init.lua
@@ -130,25 +130,25 @@ local driver_template = {
     }
   },
   sub_drivers = {
-    require("eaton-accessory-dimmer"),
-    require("inovelli-LED"),
-    require("dawon-smart-plug"),
-    require("inovelli-2-channel-smart-plug"),
-    require("zwave-dual-switch"),
-    require("eaton-anyplace-switch"),
-    require("fibaro-wall-plug-us"),
-    require("dawon-wall-smart-switch"),
-    require("zooz-power-strip"),
-    require("aeon-smart-strip"),
-    require("qubino-switches"),
-    require("fibaro-double-switch"),
-    require("fibaro-single-switch"),
-    require("eaton-5-scene-keypad"),
-    require("ecolink-switch"),
-    require("multi-metering-switch"),
-    require("zooz-zen-30-dimmer-relay"),
     require("multichannel-device"),
-    require("aeotec-smart-switch")
+    require("eaton-accessory-dimmer.can_handle"),
+    -- require("inovelli-LED"),
+    -- require("dawon-smart-plug"),
+    -- require("inovelli-2-channel-smart-plug"),
+    -- require("zwave-dual-switch"),
+    -- require("eaton-anyplace-switch"),
+    -- require("fibaro-wall-plug-us"),
+    -- require("dawon-wall-smart-switch"),
+    -- require("zooz-power-strip"),
+    -- require("aeon-smart-strip"),
+    -- require("qubino-switches"),
+    -- require("fibaro-double-switch"),
+    -- require("fibaro-single-switch"),
+    -- require("eaton-5-scene-keypad"),
+    -- require("ecolink-switch"),
+    -- require("multi-metering-switch"),
+    -- require("zooz-zen-30-dimmer-relay"),
+    -- require("aeotec-smart-switch")
   },
   lifecycle_handlers = {
     init = device_init,
